Fixes ClassCastException with generics and Integer as generic type
Using generics with Integer as the generic type can cause a "ClassCastException: java.lang.Long cannot be cast to java.lang.Integer".

This problem occurs randomly due to the behavior of the method "Class<?>.getDeclaredMethods" -> "The elements in the array returned are not sorted and are not in any particular order.".
In the case the exception occurs the generated bridge method is used.

A simple isBridge check in the methods "isAccessor" and "isMutator" resolves the problem.